% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBI.R
\name{BI}
\alias{BI}
\title{Standardized Ball Imformation}
\usage{
BI(x, y, R = 0, seed = 2015, weight = FALSE)
}
\arguments{
\item{x}{a numeric matirx.}

\item{y}{a numeric matirx.}

\item{R}{When R=0, it will return Ball information}

\item{seed}{Not need}

\item{weight}{Not need}
}
\description{
Calculate Ball Information statistic
}
\examples{

n = 100
x = rnorm(n)
y = rnorm(n)
SBI(x, y)

}

