\name{GSE}
\alias{GSE}
\docType{data}
\title{
GSE5327
}
\description{GSE5327 includes 58 samples that contain the expression of
22,283 genes, and characteristics of the sample such as metastasis status and metastasisfree
time.}
\usage{data(GSE)}
\details{It is available in NCBI website with the keyword "GSE5327""}
\examples{
data(GSE)
}
\keyword{datasets}
