% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBI.R
\name{SBI}
\alias{SBI}
\title{Standardized Ball Imformation}
\usage{
SBI(x, y)
}
\arguments{
\item{x}{a numeric matirx.}

\item{y}{a numeric matirx.}
}
\description{
Calculate Ball Information statistic
}
\examples{

n = 100
x = rnorm(n)
y = rnorm(n)
SBI(x, y)

}

